% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band2Raster.R
\name{band2Raster}
\alias{band2Raster}
\title{\code{band2Raster} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree. It then writes the resulting files in h5 format for machine learning input}
\usage{
band2Raster(file, band, noDataValue, xMin, yMin, res, crs)
}
\arguments{
\item{file:}{the hdf file}

\item{band:}{the band you want to process}
}
\value{
returns: a matrix containing the reflectance data for the specific band
}
\description{
\code{band2Raster} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree. It then writes the resulting files in h5 format for machine learning input
}
