% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tile.R
\name{check_tile}
\alias{check_tile}
\title{Check if tile can be used for training}
\usage{
check_tile(itcs_path, lidar_path, rgb_dir)
}
\arguments{
\item{itcs_path}{path to itcs}

\item{lidar_path}{Character. path to lidar data directory}

\item{rgb_dir}{Character. path to camera tile}

\item{cores}{Numeric. Available cores to process data}
}
\value{
Writes training crops to file
}
\description{
\code{check_tile} ensures that training tiles 1) Do not overlap with ground truth, and 2) have both lidar and rgb data
}
