% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_PR_site.R
\name{compute_PR_site}
\alias{compute_PR_site}
\title{Site wrapper for calculating recall and precision from IoU overlap for a NEON site.}
\usage{
compute_PR_site(site, algorithm = "Silva", threshold = 0.5, plot = FALSE,
  save = FALSE, bounding_box = FALSE)
}
\arguments{
\item{algorithm}{Character. "Silva","Li", "Dalponte", see lidr::lastrees}

\item{lidar_file}{Character. Path to lidar .laz on disk}

\item{Bounding}{Box Character. Whether to return the convex hulls or bounding box of predicted clouds}
}
\value{
a dataframe of objects for all plots in a site
}
\description{
Site wrapper for calculating recall and precision from IoU overlap for a NEON site.
}
