% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePR.R
\name{computePR}
\alias{computePR}
\title{Wrapper for calculating recall and precision from IoU overlap for a NEON site.
\code{computeMAP} computes an lidar based segmentation, assigns polygons to closest match and calculates recall statistic}
\usage{
computePR(lidar_file, plot_data, algorithm = "Silva", threshold = 0.5,
  plot = TRUE, save = FALSE, bounding_box = FALSE)
}
\arguments{
\item{lidar_file}{Character. Path to lidar .laz on disk}

\item{plot_data}{Data.Frame ground truth annotations from parse_xml().}

\item{algorithm}{Character. "Silva","Li", "Dalponte", see lidr::lastrees}
}
\value{
A dataframe of matched objects
}
\description{
Wrapper for calculating recall and precision from IoU overlap for a NEON site.
\code{computeMAP} computes an lidar based segmentation, assigns polygons to closest match and calculates recall statistic
}
