% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus}
\alias{consensus}
\title{Majority voting for individual tree crown segmentation based on multiple algorithms}
\usage{
consensus(ptlist, method = "majority", debug = F)
}
\arguments{
\item{ptlist}{A list of lidar clouds read in by lidR package, processed using \code{\link[lidR]{lastrees}}}

\item{method}{Character. A cluster voting scheme. Currently implemented "majority" or "kmodes" see \code{\link[diceR]{majority_voting}}  processed using \code{\link[lidR]{lastrees}}}
}
\value{
A las object with treeID field updated based on ensemble individual tree segmentation. A new column "consensus" is the consensus treeID.
}
\description{
\code{consensus} assigns each point in a lidR cloud to a treeID based on multiple segmentation algorithms
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
tile = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0"
chm=canopy_model(tile)
ptlist<-list()
ptlist[["watershed"]] <- segment_trees(tile,algorithm="watershed",chm=chm)
ptlist[["silva2016"]]<- segment_trees(tile,algorithm="silva2016",chm=chm)
consensus(ptlist)
}
