% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_names.R
\name{convert_names}
\alias{convert_names}
\title{Get corresponding tile across data types}
\usage{
convert_names(from, to, lidar = NULL, rgb = NULL, hyperspectral = NULL,
  site = "OSBS")
}
\arguments{
\item{from}{current sensor}

\item{to}{destination sensor}

\item{lidar}{Character. Lidar pathname}

\item{rgb}{Character. RGB pathname}

\item{hyperspectral}{hyerspectral pathname}
}
\description{
\code{convert_names} uses the geocode from each NEON data product to find the corresponding tile in another sensor
}
