% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull.R
\name{convex_hull}
\alias{convex_hull}
\title{Find convex hull that outlines an individual tree}
\usage{
convex_hull(x, plot = FALSE)
}
\arguments{
\item{x}{A two column matrix with column names "X" and "Y"}

\item{plot}{Whether to plot the results for visualization}
}
\value{
A \code{\link[sp]{SpatialPolygons}} object containing a convex hull based on input points.
}
\description{
\code{convex_hull} finds the outer hull of a set of points.
}
\examples{
Not usually called directly, see \\code{\\link{get_convex_hulls}}
}
