% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateFromAzimuth.R
\name{CoordinateFromAzimuth}
\alias{CoordinateFromAzimuth}
\title{Extract geographic coordinates from plot center measurements}
\usage{
CoordinateFromAzimuth(dista, angle, xcord, ycord)
}
\description{
\code{CoordinateFromAzimuth} returns a canopy height raster from a lidR point cloud
}
