% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_CHM_plots.R
\name{crop_CHM_plots}
\alias{crop_CHM_plots}
\title{Clip CHM Data Based on Neon Plots}
\usage{
crop_CHM_plots(site_name = "TEAK", year = "2018")
}
\arguments{
\item{site_name}{NEON site abbreviation (e.g. "HARV")}
}
\value{
Saved tif files for each plot
}
\description{
\code{crop_CHM_plots} overlays the polygons of the NEON plots with the derived CHM image
}
