% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_hyperspectral_plots.R
\name{crop_hyperspectral_plots}
\alias{crop_hyperspectral_plots}
\title{Clip Hyperspectral Data Based on extent of NEON Tower Plots}
\usage{
crop_hyperspectral_plots(siteID = "TEAK", year = "2018",
  false_color = FALSE)
}
\arguments{
\item{siteID}{NEON site abbreviation (e.g. "HARV")}
}
\value{
Saved tif files for each plot
}
\description{
Clip Hyperspectral Data Based on extent of NEON Tower Plots
}
