% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_lidar_plots.R
\name{crop_lidar_plots}
\alias{crop_lidar_plots}
\title{Clip Lidar Data Based on Neon Plots}
\usage{
crop_lidar_plots(siteID = "TEAK")
}
\arguments{
\item{siteID}{NEON site abbreviation (e.g. "HARV")}
}
\value{
Saved tif files for each plot
}
\description{
\code{crop_lidar_plots} overlays the polygons of the NEON plots with the lidar airborne data
}
