% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_rgb_plots.R
\name{crop_rgb_plots}
\alias{crop_rgb_plots}
\title{Clip RGB Data Based on Neon Plots}
\usage{
crop_rgb_plots(siteID = "SJER", year = "2018")
}
\arguments{
\item{siteID}{NEON site abbreviation (e.g. "HARV")}
}
\value{
Saved tif files for each plot
}
\description{
\code{crop_rgb_plots} overlays the polygons of the NEON plots with the RGB airborne data
}
