% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_target_CHM.R
\name{crop_target_CHM}
\alias{crop_target_CHM}
\title{Clip CHM Data Based on extent of RGB file}
\usage{
crop_target_CHM(siteID = "TEAK", rgb_filename, year = "2019",
  tif_base_dir = "/orange/ewhite/NeonData",
  save_base_dir = "/orange/ewhite/b.weinstein/NEON")
}
\arguments{
\item{rgb_filename}{path to a projected file to extract coordinates}

\item{year}{year to match in CHM files}

\item{tif_base_dir}{where to search for CHM files}

\item{save_dir}{base path to save cropped files}
}
\value{
Saved tif files for each plot
}
\description{
Clip CHM Data Based on extent of RGB file
}
