% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalponte2016.R
\name{dalponte2016}
\alias{dalponte2016}
\title{Find tree crown polygons using the Dalponte 2016 algorithm}
\usage{
dalponte2016(path = NULL, tile = NULL, output = "all")
}
\arguments{
\item{path}{Character. A filename of a .las or .laz file to be read in by the lidR package, see \code{\link[lidR]{readLAS}}}

\item{tile}{Character. Optionally a lidR object in memory}
}
\description{
\code{dalponte2016} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- dalponte2016(path=LASfile)
}
