% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_training_benchmark.R
\name{detection_training_benchmark}
\alias{detection_training_benchmark}
\title{Generate detection network training files}
\usage{
detection_training_benchmark(path, silva_cr_factor, silva_exclusion, rgb_tiles,
  save_dir = "Results/detection_boxes/")
}
\arguments{
\item{path}{Character Location of the lidar tile to process.}

\item{silva_cr_factor}{see lidR::lastrees}

\item{silva_exclusion}{see lidR::lastrees}

\item{save_dir}{Where to save the boxes, a folder for the year will be added to relative to this base dir.}
}
\value{
A .csv file is written to disk containing detections
}
\description{
\code{detection training} is the pipeline for generating training data for a detection based nueral net (RCNN).
}
