% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Wrapper for calculating overlap among predictiong and ground truth polygons for lidar-based tree segmentation methods}
\usage{
evaluate(ground_truth, algorithm = "silva", path_to_tiles = NULL,
  compute_consensus = F, extra = F)
}
\arguments{
\item{ground_truth}{SpatialPolygonDataFrame of ground truth polygons}

\item{algorithm}{Character. A vector of lidar unsupervised classification algorithm(s). Currently "silva","dalponte","li" and "watershed" are implemented. See \code{\link[lidR]{lastrees}}}

\item{path_to_tiles}{Character. Location of lidar tiles on system.}

\item{compute_consensus}{Logical. Generate a consensus from selected methods, see \code{\link{consensus}}.}

\item{extra}{Logical. Return a list of segmented lidar tiles, predicted convex hull polygons, and the calculated evaluation statistics.}
}
\value{
dataframe of the jaccard overlap among polygon pairs for each selected method. If extra=T, \code{evaluate} will return a list object of results, predicted polygons, as well as output lidR tiles. See e.g. \link{\code{silva2016}}
}
\description{
\code{evaluate} computes an lidar based segmentation, assigns polygons to closest match and calculates jaccard stat
}
