% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateNeon.R
\name{evaluateNeon}
\alias{evaluateNeon}
\title{Wrapper for calculating recall statistics for NEON tree centroids.}
\usage{
evaluateNeon(trees, plotID, algorithm = "silva", path_to_tiles = NULL,
  extra = F, plot_results = F, basemap = "", epsg_numeric = 32611)
}
\arguments{
\item{algorithm}{Character. A vector of lidar unsupervised classification algorithm(s). Currently "silva","dalponte","li" and "watershed" are implemented. See \code{\link[lidR]{lastrees}}}

\item{path_to_tiles}{Character. Location of lidar tiles on system.}

\item{extra}{Logical. Return a list of segmented lidar tiles, predicted convex hull polygons, and the calculated evaluation statistics.}

\item{plot_results}{Logical. Generate a plot of ground truth and predicted polygons}

\item{basemap}{Character Directory of rgb images for basemap}

\item{ground_truth}{SpatialPolygonDataFrame of ground truth polygons}
}
\value{
dataframe of the jaccard overlap among polygon pairs for each selected method. If extra=T, \code{evaluate} will return a list object of results, predicted polygons, as well as output lidR tiles. See e.g. \code{\link{silva2016}}
}
\description{
\code{evaluateNeon} computes an lidar based segmentation, assigns polygons to closest match and calculates recall statistic
}
