% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_dist_to_utm.R
\name{from_dist_to_utm}
\alias{from_dist_to_utm}
\title{Convert coordinates to utms
\code{from_dist_to_utm} converts azimuth distance to utm coodinates}
\usage{
from_dist_to_utm(dista, angle, xcord, ycord)
}
\arguments{
\item{dista}{Distance from point}

\item{angle}{Angle from point}

\item{xcord}{X distance}

\item{ycord}{Y distance}
}
\value{
A stacked file in /data/ with desired product across sites
}
\description{
Convert coordinates to utms
\code{from_dist_to_utm} converts azimuth distance to utm coodinates
}
