% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_training.R
\name{generate_training}
\alias{generate_training}
\title{Generate training data from lidar tiles}
\usage{
generate_training(las = NULL, algorithm = "silva", cores = NULL)
}
\arguments{
\item{las}{Character or Vector. Path to lidar tiles on system to be processed. May be a single tile or vector of tiles.}

\item{algorithm}{Character. A vector of lidar unsupervised classification algorithm(s). Currently "silva","dalponte","li" and "watershed" are implemented. See \code{\link[lidR]{lastrees}}}
}
\value{
A nested list of lidR tiles equal to the length of path_to_tiles. Each list will have a list of segmented .las for each tree.
}
\description{
\code{generate_training} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree. It then writes the resulting files in h5 format for machine learning input
}
