% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box.R
\name{get_box}
\alias{get_box}
\title{Get a bounding box from a data.frame}
\usage{
get_box(x, expand)
}
\arguments{
\item{x}{data.frame Containing a column of X and Y coordinates}

\item{expand.}{Numeric. Multiply box size by a factor.}
}
\value{
A list of \code{\link[sp]{SpatialPolygonsDataFrame}} objects containing a bounding box for each defined tree.
}
\description{
\code{get_box} Takes a data.frame of x,y coordinates and returns the extent
}
