% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_convex_hulls.R
\name{get_convex_hulls}
\alias{get_convex_hulls}
\title{Generate a SpatialPolygonsDataFrame of convex hulls for all segmented trees}
\usage{
get_convex_hulls(tile, ID)
}
\arguments{
\item{tile}{A lidR point cloud which has been processed by \code{\link[lidR]{lastrees}}, see \code{\link{segment_trees}}}

\item{ID}{The name of the ID field that deliniates individual trees.}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}} object containing a convex hull for each defined tree.
}
\description{
\code{get_convex_hulls} Iterates through a lidR .las point cloud and returns convex hulls for all identified trees.
}
\examples{
Read in tile
tile=readLAS("../tests/data/NEON_D03_OSBS_DP1_404000_3284000_classified_point_cloud.laz")
#Generate ground model
ground_model(tile)
#Generate canopy model
chm=canopy_model(tile)

#remove ground points? Classification == 2?
tile<-tile \%>\% lasfilter(!Classification==2)

#Compute unsupervised classification method
system.time(silva2016<-segment_ITC(tile,algorithm = "silva2016",chm=chm))
#create polygons
silva_convex<-get_convex_hull(silva2016,silva2016@data$treeID)

}
