% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile_filename_multiple.R
\name{get_tile_filname_multiple}
\alias{get_tile_filname_multiple}
\title{Get correpsonding .laz tile from file based on the ground truth polygon}
\usage{
get_tile_filname_multiple(polygon, basename = "NEON_D03_OSBS_DP1_",
  extension = "_classified_point_cloud.laz")
}
\arguments{
\item{polygon}{The SpatialPolygon file for the ground truth polygon}

\item{basename}{The directory containing the .laz files.}
}
\value{
A file path
}
\description{
\code{get_tile_filname} returns the path on file based on extent of the polygon
}
\examples{
Not usually called directly by user.
}
