% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_TOS_data}
\alias{get_TOS_data}
\title{Download terrestrial data from NEON
\code{get_TOS_data} looks up which sites have data for a given terrestrial product.}
\usage{
get_TOS_data(prd = NULL)
}
\arguments{
\item{prd}{Product ID number, e.g. DP1.10098.001 (veg structure ) is 10098}
}
\value{
A stacked file in /data/ with desired product across sites
}
\description{
Download terrestrial data from NEON
\code{get_TOS_data} looks up which sites have data for a given terrestrial product.
}
