% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vegetation_structure.R
\name{get_vegetation_structure}
\alias{get_vegetation_structure}
\title{Organize terrestrial data from NEON
\code{get_vegetatation_structure} cleans the downloaded NEON data}
\usage{
get_vegetation_structure(path)
}
\arguments{
\item{prd}{Product ID number, e.g. DP1.10098.001 (veg structure ) is 10098}
}
\value{
A stacked file in /data/ with desired product across sites
}
\description{
Organize terrestrial data from NEON
\code{get_vegetatation_structure} cleans the downloaded NEON data
}
