% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_to_rgb.R
\name{h5_to_rgb}
\alias{h5_to_rgb}
\title{\code{h5_to_rgb} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree. It then writes the resulting files in h5 format for machine learning input}
\usage{
h5_to_rgb(file_path, write = F, outpath = NULL, bands = list(58, 34, 19))
}
\arguments{
\item{file_path}{Character. path to h5 file.}

\item{outpath}{Character. fullpath to write RGB path.}

\item{bands}{List Bands to combine into stack}
}
\value{
returns: a RGB raster stack
}
\description{
\code{h5_to_rgb} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree. It then writes the resulting files in h5 format for machine learning input
}
