% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgcat.R
\name{imgcat}
\alias{imgcat}
\title{Plot result in iTerm}
\usage{
imgcat(plot_command)
}
\arguments{
\item{plot_command}{a plot function to be viewed in the terminal}
}
\description{
\code{imgcat} assigns each point in a lidR cloud to a treeID.
}
