% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li2012.R
\name{li2012}
\alias{li2012}
\title{Find tree crown polygons using the Li (2012) algorithm}
\usage{
li2012(path = NULL, tile = NULL, output = "all")
}
\arguments{
\item{path}{Character. A filename of a .las or .laz file to be read in by the lidR package, see \code{\link[lidR]{readLAS}}}

\item{tile}{Character. Optionally a lidR object in memory}

\item{output}{Character. 'Tile', "convex_hull","all"}
}
\value{
\preformatted{
'tile' : .las tile with segmented trees in the treeID column
'convex_hull': a \code{\link[sp]{SpatialPolygonsDataFrame}} object with tree crown polygons
'all': Both a tile and convex hull output in a named list.
}
}
\description{
\code{li2012} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- li2012(path=LASfile)
}
