% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_samples.R
\name{negative_samples}
\alias{negative_samples}
\title{Negative sampling of a tile based on observed trees}
\usage{
negative_samples(boxes, path_las)
}
\arguments{
\item{path_las}{A tile to define the bounds of sampling}

\item{boxes.}{A dataframe output from \code{training_crops}}
}
\description{
\code{negative_samples} Negative training samples are produced by randomly placing boxes along the tiles. The size of each box, and the number of total tiles, match the number of trees found in the tile.
}
