% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_xml.R
\name{parse_xml}
\alias{parse_xml}
\title{Site wrapper for calculating recall and precision from IoU overlap for a NEON site.}
\usage{
parse_xml(site)
}
\arguments{
\item{site}{Character. NEON Site code}
}
\value{
a data.frame object with the spatial polygons of the ground truth annotations
}
\description{
Site wrapper for calculating recall and precision from IoU overlap for a NEON site.
}
