% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_polygons.R
\name{plot_3d_polygons}
\alias{plot_3d_polygons}
\title{Wrapper for plotting a set of polygons on to a 3d lidar plot}
\usage{
plot_3d_polygons(spdf, z = 0)
}
\arguments{
\item{spdf}{A SpatialPolygonDataFrame or polygons with length 1}

\item{z}{The height to plot polygons.}
}
\value{
NULL - plot invoked by side effect.
}
\description{
\code{plot_3d_polygons} creates 3d polygons from a SpatialPolygonsDataFrame
}
