% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_field_data.R
\name{plot_field_data}
\alias{plot_field_data}
\title{Plot field data on basemaps}
\usage{
plot_field_data(site = "MLBS",
  image_dir = "/Users/Ben/Documents/NeonTreeEvaluation/evaluation")
}
\arguments{
\item{image_dir}{path to the image directory (e.g NeonTreeEvaluation/)}

\item{siteID}{NEON site abbreviation (e.g. "HARV")}
}
\value{
Saved tif files for each plot
}
\description{
\code{plot_field data} overlays the woody vegetation structure NEON data on the airborne RGB mosiac.
}
