% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_overlap.R
\name{polygon_overlap}
\alias{polygon_overlap}
\title{Find area of overlap among sets of polygons}
\usage{
polygon_overlap(pol, prediction)
}
\arguments{
\item{pol}{A ground truth polygon}

\item{prediction}{Whether to plot the results for visualization}
}
\value{
A data frame with the crown ID, the prediction ID and the area of overlap.
}
\description{
\code{convex_hull} finds the outer hull of a set of points.
}
\examples{

Not usually called directly by user
}
