% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dalponte2016.R
\name{run_dalponte2016}
\alias{run_dalponte2016}
\title{Find tree crown polygons using the Dalponte 2016 algorithm}
\usage{
run_dalponte2016(path = NULL, tile = NULL, output = "all", epsg_numeric)
}
\description{
\code{dalponte2016} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- run_dalponte2016(path=LASfile)
}
