% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_detection.R
\name{run_detection}
\alias{run_detection}
\title{Run detection training for a tile in a site}
\usage{
run_detection(lidar_file, site, rgb_dir, year, silva_cr_factor, silva_exclusion)
}
\arguments{
\item{lidar_file}{Character. path to lidar file}

\item{site}{Four letter NEON abbreviation}

\item{rgb_dir}{Character. Directory of rgb files to check}

\item{silva_cr_factor}{see ?run_silva}

\item{silva_exclusion}{see ?run_silva}
}
\value{
NULL; CSV files written to file
}
\description{
\code{run_detection} is a wrapper for detection_training for checking a file exists and optionally overwriting
}
