% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_silva2016.R
\name{run_silva2016}
\alias{run_silva2016}
\title{Find tree crown polygons using the Silva 2016 algorithm}
\usage{
run_silva2016(path = NULL, tile = NULL, output = c("all"), epsg_numeric,
  max_cr_factor = 0.7, exclusion = 0.3)
}
\description{
\code{run_silva2016} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- run_silva2016(path=LASfile)
}
