% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_watershed.R
\name{run_watershed}
\alias{run_watershed}
\title{Find tree crown polygons using a simple watershed algorithm}
\usage{
run_watershed(path = NULL, tile = NULL, output = "all", epsg_numeric)
}
\description{
\code{run_watershed} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- watershed(path=LASfile)
}
