% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silva2016.R
\name{silva2016}
\alias{silva2016}
\title{Find tree crown polygons using the Silva 2016 algorithm}
\usage{
silva2016(path = NULL, tile = NULL, output = c("all"))
}
\arguments{
\item{path}{Character. A filename of a .las or .laz file to be read in by the lidR package, see \code{\link[lidR]{readLAS}}}

\item{tile}{Character. Optionally a lidR object in memory}

\item{output}{Character. 'Tile', "convex_hull","all"}
}
\description{
\code{silva2016} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- silva2016(path=LASfile)
}
