% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_crops.R
\name{training_crops}
\alias{training_crops}
\title{Generate bounding boxes from lidar tiles}
\usage{
training_crops(path_las = NULL, algorithm = "silva", cores = NULL,
  create_negatives = T, expand = 1)
}
\arguments{
\item{create_negatives}{Should negative samples be generated. see details}

\item{expand}{Numeric. Factor to enlarge (or decrease) bounding box size}

\item{path_rgb}{Character. file path of rgb image}
}
\description{
\code{training_crops} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree. It then writes the resulting files in h5 format for machine learning input
}
\details{
Negative training samples are produced by randomly placing boxes along the tiles. The size of each box, and the number of total tiles, match the number of trees found in the tile.
}
