% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeClusters.R
\name{treeClusters}
\alias{treeClusters}
\title{Find clusters of elevation on the canopy height model}
\usage{
treeClusters(path, threshold = 15, res = 3, expand = 0)
}
\arguments{
\item{path}{Character. file path to lidar image}

\item{threshold}{Minimum height of the tree clusters.}

\item{res}{Numeric. Resolution of the canopy height model}

\item{expand}{Numeric. Meters to expand the bounding box to pad the edges.}
}
\value{
A data.frame of bounding boxes for each tree object.
}
\description{
\code{treeClusters} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree. It then writes the resulting files in h5 format for machine learning input
}
