% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{watershed}
\alias{watershed}
\title{Find tree crown polygons using a simple watershed algorithm}
\usage{
watershed(path = NULL, tile = NULL, output = "all")
}
\arguments{
\item{path}{Character. A filename of a .las or .laz file to be read in by the lidR package, see \code{\link[lidR]{readLAS}}}

\item{tile}{Character. Optionally a lidR object in memory}

\item{output}{Character. 'Tile', "convex_hull","all"}
}
\description{
\code{watershed} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- watershed(path=LASfile)
}
