% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalponte2016.R
\name{dalponte2016}
\alias{dalponte2016}
\title{Find tree crown polygons using the Dalponte 2016 algorithm}
\usage{
dalponte2016(path = NULL, tile = NULL, extra = F)
}
\arguments{
\item{path}{A filename of a .las or .laz file to be read in by the lidR package}

\item{tile}{Optionally a lidR object in memory}

\item{extra}{Output both the tile and the convex polygons}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}} object with tree crown polygons
}
\description{
\code{dalponte2016} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- dalponte2016(path=LASfile)
}
