% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_training.R
\name{detection_training}
\alias{detection_training}
\title{Generate detection network training files}
\usage{
detection_training(path)
}
\arguments{
\item{path}{Character Location of the lidar tile to process.}
}
\value{
A .csv file is written to disk containing detections
}
\description{
\code{detection training} is the pipeline for generating training data for a detection based nueral net (RCNN).
}
