% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Wrapper for calculating overlap among predictiong and ground truth polygons for lidar-based tree segmentation methods}
\usage{
evaluate(ground_truth, algorithm = "silva", path_to_tiles = NULL)
}
\arguments{
\item{ground_truth}{A SpatialPolygonDataFrame of ground truth polygons}

\item{algorithm}{A lidar unsupervised classification algorithm. Currently only "silva" is available.}

\item{path_to_tiles}{Location of lidar tiles on system.}
}
\value{
dataframe of the jaccard overlap among polygon pairs
}
\description{
\code{evaluate} computes an lidar based segmentation, assigns polygons to closest match and calculates jaccard stat
}
