% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_all.R
\name{evaluate_all}
\alias{evaluate_all}
\title{Wrapper among tiles for calculating overlap among predictiong and ground truth polygons for lidar-based tree segmentation methods}
\usage{
evaluate_all(itcs, algorithm = "silva", path_to_tiles = NULL)
}
\arguments{
\item{algorithm}{A lidar unsupervised classification algorithm. Currently only "silva" is available.}

\item{path_to_tiles}{Location of lidar tiles on system.}

\item{ground_truth}{A SpatialPolygonDataFrame of ground truth polygons}
}
\value{
dataframe of the jaccard overlap among polygon pairs
}
\description{
\code{evaluate_all} extends evaluate to multiple tiles
}
