% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_all.R
\name{evaluate_all}
\alias{evaluate_all}
\title{Wrapper among tiles for calculating overlap among predicted and ground truth polygons for lidar-based tree segmentation methods
\code{evaluate_all} extends evaluate to multiple tiles, optionally run in parallel.}
\usage{
evaluate_all(itcs, algorithm = "silva", path_to_tiles = NULL,
  cores = NULL, extra = F, plot_results = F, basemap = "")
}
\arguments{
\item{itcs}{A SpatialPolygonDataFrame of ground truth polygons}

\item{algorithm}{Character. A vector of lidar unsupervised classification algorithm(s). Currently "silva","dalponte","li" and "watershed" are implemented. See \code{\link[lidR]{lastrees}}}

\item{path_to_tiles}{Character. Location of lidar tiles on system.}

\item{cores}{Optional parameter to specify the number of cores to parallelize using \code{\link[foreach]{foreach}}}

\item{extra}{Logical. Return a list of segmented lidar tiles, predicted convex hull polygons, and the calculated evaluation statistics.}

\item{plot_results}{Logical. Generate a plot of ground truth and predicted polygons}

\item{basemap}{Character Directory of rgb images for basemap}
}
\value{
dataframe of the jaccard overlap among polygon pairs
}
\description{
Wrapper among tiles for calculating overlap among predicted and ground truth polygons for lidar-based tree segmentation methods
\code{evaluate_all} extends evaluate to multiple tiles, optionally run in parallel.
}
