% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trees.R
\name{extract_trees}
\alias{extract_trees}
\title{Extract individual las files for segmented trees}
\usage{
extract_trees(path_to_tiles = NULL, algorithm = "silva",
  compute_consensus = F, cores = NULL)
}
\arguments{
\item{path_to_tiles}{Character or Vector. Location of lidar tiles on system. May be a single tile or vector of tiles.}

\item{algorithm}{Character. A vector of lidar unsupervised classification algorithm(s). Currently "silva","dalponte","li" and "watershed" are implemented. See \code{\link[lidR]{lastrees}}}

\item{compute_consensus}{Logical. Generate a consensus from selected methods, see \code{\link{consensus}}.}
}
\value{
a list of lidR tiles.
}
\description{
\code{extra_trees} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree.
}
