% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_training.R
\name{generate_training}
\alias{generate_training}
\title{Generate training crop data from lidar tiles}
\usage{
generate_training(lidar = NULL, rgb = NULL, hyperspec = NULL,
  algorithm = "silva", cores = NULL, outdir = NULL, write = T)
}
\arguments{
\item{lidar}{Character. list of lidar paths}

\item{rgb}{Character. list of rgb image paths}

\item{hyperspec}{Character. list of hyperspec paths}

\item{cores}{Numeric. Available cores to process data}

\item{outdir}{Character. Location on disk to write training crops}
}
\value{
Writes training crops to file
}
\description{
\code{training_crops} computes an lidar-based segmentation, based on multiple available methods, and splits the results into individual las files for each predicted tree. It then writes the resulting files in h5 format for machine learning input
}
