% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_model.R
\name{ground_model}
\alias{ground_model}
\title{Generate a ground model from raw las}
\usage{
ground_model(las)
}
\arguments{
\item{las}{A lidar cloud read in by lidR package}
}
\value{
NA. By default operation is done in place to reduce memory.
}
\description{
\code{ground_model} computes a ground model and normalizes the input point cloud such that ground points are at zero elevation
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
tile = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0"
ground_model(tile)

}
