% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li2012.R
\name{li2012}
\alias{li2012}
\title{Find tree crown polygons using the Li (2012) algorithm}
\usage{
li2012(path = NULL, tile = NULL, extra = F)
}
\arguments{
\item{path}{A filename of a .las or .laz file to be read in by the lidR package}

\item{tile}{Optionally a lidR object in memory}

\item{extra}{Output both the tile and the convex polygons}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}} object with tree crown polygons
}
\description{
\code{li2012} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- li2012(path=LASfile)
}
