% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_trees.R
\name{segment_trees}
\alias{segment_trees}
\title{Find individual tree crowns using normalized las and a canopy height model}
\usage{
segment_trees(las, algorithm = "watershed", chm = chm)
}
\arguments{
\item{las}{A lidar cloud read in by lidR package}

\item{algorithm}{a segmentation method, see \code{\link[lidR]{lastrees}}}

\item{chm}{a canopy height model see \code{\link{canopy_model}}}
}
\value{
A las object with treeID field updated based on individual tree segmentation.
}
\description{
\code{segment_trees} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
tile = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0"
chm=canopy_model(tile)
treelas=segment_trees(tile,algorithm="watershed",chm=chm)
}
