% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silva2016.R
\name{silva2016}
\alias{silva2016}
\title{Find tree crown polygons using the Silva 2016 algorithm}
\usage{
silva2016(path)
}
\arguments{
\item{path}{A filename of a .las or .laz file to be read in by the lidR package}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}} object with tree crown polygons
}
\description{
\code{silva2016} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
tile = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")
convex_hulls <- silva2016(tile)
}
