% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{watershed}
\alias{watershed}
\title{Find tree crown polygons using a simple watershed algorithm}
\usage{
watershed(path = NULL, tile = NULL, extra = F)
}
\arguments{
\item{path}{A filename of a .las or .laz file to be read in by the lidR package}

\item{tile}{Optionally a lidR object in memory}

\item{extra}{Output both the tile and the convex polygons}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}} object with tree crown polygons
}
\description{
\code{watershed} assigns each point in a lidR cloud to a treeID.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
convex_hulls <- watershed(path=LASfile)
}
